Expression Sets

The ExpressionSet type is a container for storing gene expression data. Following the Bioconductor ExpressionSet class, it contains a matrix of expression values, a matrix of feature data, and a matrix of sample data.

Constructor

ExpressionData.ExpressionSet — Type

ExpressionSet

An ExpressionSet object is a container for storing gene expression data, along with associated metadata. It follows the ExpressionSet class from the R package from Bioconductor: Biobase.

See also

MIAME feature_names sample_names expression_values feature_data phenotype_data experiment_data annotation

Accessors

ExpressionData.expression_values — Function

expression_values(eset::ExpressionSet)::DataFrame

Extracts the expression values from an ExpressionSet. The feature names are included as a column.

By passing the type Matrix as the first argument, the function will return the expression values as a matrix.

ExpressionData.feature_data — Function

feature_data(eset::ExpressionSet)::DataFrame

Extracts the feature data from an ExpressionSet.

ExpressionData.phenotype_data — Function

phenotype_data(eset::ExpressionSet)::DataFrame

Extracts the phenotype data from an ExpressionSet.

ExpressionData.experiment_data — Function

experiment_data(eset::ExpressionSet)::MIAME

Extracts the experiment data from an ExpressionSet.

See also

MIAME

ExpressionData.annotation — Function

annotation(eset::ExpressionSet)::Symbol

Extracts the annotation from an ExpressionSet.

ExpressionData.sample_names — Function

sample_names(eset::ExpressionSet)::Vector{String}

Extracts the sample names from an ExpressionSet.

ExpressionData.feature_names — Function

feature_names(eset::ExpressionSet)::Vector{String}

Extracts the feature names from an ExpressionSet.