Saving and loading data

The ExpressionData package provides functions for saving and loading gene expression data in various formats.

Saving

The save_eset function saves gene expression data to a file in a specified format.

ExpressionData.save_eset — Function

save_eset(eset::ExpressionSet, file::AbstractString)

Saves an ExpressionSet object to a file. The file is saved in a serialized format. To load the object back, use load_eset.

See also

load_eset

Loading

The load_eset function loads gene expression data from a file saved through save_eset. Support reading from RDS files is also provided.

ExpressionData.load_eset — Function

load_eset(file::AbstractString)::ExpressionSet

Loads an ExpressionSet object from a file. The file should be in a serialized format. To save an object, use save_eset.

See also

save_eset