MIAME

The MIAME struct provides functions for working with Minimum Information About a Microarray Experiment (MIAME) data.

Constructor

ExpressionData.MIAME — Type

MIAME

An MIAME object is a container for storing metadata associated with an experiment. It follows the MIAME class from the R package from Bioconductor: Biobase.

See also

abstract info hybridizations norm_controls other preprocessing pub_med_id

Accessors

ExpressionData.abstract — Function

abstract(m::MIAME)::String

Extracts the abstract from an MIAME object.

ExpressionData.info — Function

info(m::MIAME)::NamedTuple

Returns a NamedTuple with the name, lab, contact, title, and url fields of an MIAME object. Similar to the expinfo function in the R package from Bioconductor: Biobase.

ExpressionData.hybridizations — Function

hybridizations(m::MIAME)::Vector{String}

Extracts the hybridizations from an MIAME object.

ExpressionData.norm_controls — Function

norm_controls(m::MIAME)::Vector{String}

Extracts the normalization controls from an MIAME object such as house keeping genes.

ExpressionData.other — Function

other(m::MIAME)::Dict{Symbol,String}

Extracts the other metadata from an MIAME object. This can include any additional information that is not covered by the other fields.

ExpressionData.preprocessing — Function

preprocessing(m::MIAME)::Vector{String}

Extracts the preprocessing steps from an MIAME object describe the steps taken to process the raw data of the experiment.

ExpressionData.pub_med_id — Function

pub_med_id(m::MIAME)::String

Extracts the pubmed id from an MIAME object.