Reference

ExpressionData.ExpressionSet — Type

ExpressionSet

An ExpressionSet object is a container for storing gene expression data, along with associated metadata. It follows the ExpressionSet class from the R package from Bioconductor: Biobase.

See also

MIAME feature_names sample_names expression_values feature_data phenotype_data experiment_data annotation

ExpressionData.MIAME — Type

MIAME

An MIAME object is a container for storing metadata associated with an experiment. It follows the MIAME class from the R package from Bioconductor: Biobase.

See also

abstract info hybridizations norm_controls other preprocessing pub_med_id

ExpressionData.abstract — Method

abstract(m::MIAME)::String

Extracts the abstract from an MIAME object.

ExpressionData.annotation — Method

annotation(eset::ExpressionSet)::Symbol

Extracts the annotation from an ExpressionSet.

ExpressionData.experiment_data — Method

experiment_data(eset::ExpressionSet)::MIAME

Extracts the experiment data from an ExpressionSet.

See also

MIAME

ExpressionData.expression_values — Method

expression_values(eset::ExpressionSet)::DataFrame

Extracts the expression values from an ExpressionSet. The feature names are included as a column.

By passing the type Matrix as the first argument, the function will return the expression values as a matrix.

ExpressionData.feature_data — Method

feature_data(eset::ExpressionSet)::DataFrame

Extracts the feature data from an ExpressionSet.

ExpressionData.feature_names — Method

feature_names(eset::ExpressionSet)::Vector{String}

Extracts the feature names from an ExpressionSet.

ExpressionData.hybridizations — Method

hybridizations(m::MIAME)::Vector{String}

Extracts the hybridizations from an MIAME object.

ExpressionData.info — Method

info(m::MIAME)::NamedTuple

Returns a NamedTuple with the name, lab, contact, title, and url fields of an MIAME object. Similar to the expinfo function in the R package from Bioconductor: Biobase.

ExpressionData.load_eset — Method

load_eset(file::AbstractString)::ExpressionSet

Loads an ExpressionSet object from a file. The file should be in a serialized format. To save an object, use save_eset.

See also

save_eset

ExpressionData.norm_controls — Method

norm_controls(m::MIAME)::Vector{String}

Extracts the normalization controls from an MIAME object such as house keeping genes.

ExpressionData.other — Method

other(m::MIAME)::Dict{Symbol,String}

Extracts the other metadata from an MIAME object. This can include any additional information that is not covered by the other fields.

ExpressionData.phenotype_data — Method

phenotype_data(eset::ExpressionSet)::DataFrame

Extracts the phenotype data from an ExpressionSet.

ExpressionData.preprocessing — Method

preprocessing(m::MIAME)::Vector{String}

Extracts the preprocessing steps from an MIAME object describe the steps taken to process the raw data of the experiment.

ExpressionData.pub_med_id — Method

pub_med_id(m::MIAME)::String

Extracts the pubmed id from an MIAME object.

ExpressionData.sample_names — Method

sample_names(eset::ExpressionSet)::Vector{String}

Extracts the sample names from an ExpressionSet.

ExpressionData.save_eset — Method

save_eset(eset::ExpressionSet, file::AbstractString)

Saves an ExpressionSet object to a file. The file is saved in a serialized format. To load the object back, use load_eset.

See also

load_eset